Reference

Index

SurvivalDistributions.ExpoDist

SurvivalDistributions.ExponentiatedWeibull

SurvivalDistributions.GeneralizedGamma

SurvivalDistributions.LogLogistic

SurvivalDistributions.PowerGeneralizedWeibull

SurvivalDistributions.censored_loglikelihood

SurvivalDistributions.censored_loglikelihood

SurvivalDistributions.cumhazard

SurvivalDistributions.cumhazard

SurvivalDistributions.hazard

SurvivalDistributions.hazard

SurvivalDistributions.loghazard

SurvivalDistributions.loghazard

SurvivalDistributions.ExpoDist — Type

ExpoDist(γ, X)

A power distribution with power γ and base distribution X<:ContinuousUnivariateDistribution is defined as the distribution that has cumulative distribution function ``F^γ`` where F was the distribution function of X.

SurvivalDistributions.ExponentiatedWeibull — Type

ExponentiatedWeibull(sigma,nu,gamma)

The ExponentiatedWeibull distribution with scale sigma, shape nu and second shape gamma has probability density function

\[f(x; parameters) = ...\]

More details and examples of usage could be provided in this docstring.

Maybe this distribution could simply be constructed from a transformation of the original Weibull ?

References:

Link to my reference so that people understand what it is

SurvivalDistributions.GeneralizedGamma — Type

GeneralizedGamma(sigma,nu,gamma)

The GeneralizedGamma distribution with scale sigma, shape nu and second shape gamma has probability density function

\[f(x; parameters) = ...\]

More details and examples of usage could be provided in this docstring.

Maybe this distribution could simply be constructed from a transformation of the original Weibull ?

References:

Link to my reference so that people understand what it is

SurvivalDistributions.LogLogistic — Type

LogLogistic(mu,sigma)

To be described...

\[f(x; parameters) = ...\]

SurvivalDistributions.PowerGeneralizedWeibull — Type

PowerGeneralizedWeibull(sigma,nu,gamma)

The PowerGeneralizedWeibull distribution with scale sigma, shape nu and second shape gamma has probability density function

\[f(x; parameters) = ...\]

More details and examples of usage could be provided in this docstring.

Maybe this distribution could simply be constructed from a transformation of the original Weibull ?

References:

Link to my reference so that people understand what it is

SurvivalDistributions.censored_loglikelihood — Method

censored_loglikelihood(X::UnivariateDistribution, t, δ)

Provide the censored logliklyhood of the distribution X at point t, with status indicatrix δ. Is if defined as

\[δ * loghazard(X,t) - cumhazard(X,t)\]

SurvivalDistributions.cumhazard — Method

loghazard(X::UnivariateDistribution, t)

Provide the cumulative hazard function of the random variable X. It is defined as $H(t) = - \ln S(t)$ where $S$ is the survival fucntion of the random variable X.

SurvivalDistributions.hazard — Method

hazard(X::UnivariateDistribution, t)

Provide the hazard fucntion of the random variable X (supposed to be a Distributions.ContinuousUnivariateDistributions) at point t. The default implementation is simply $vh(t) = \frac{f(t)}{S(t)}$ where $f$ and $S$ are the density and survival function of X.

SurvivalDistributions.loghazard — Method

loghazard(X, t)

Provide the log of the hazard fucntion of the random variable X. See hazard for the formal definition.

Bibliography

[1]

E. Stacy. A generalization of the gamma distribution. The Annals of Mathematical Statistics 33, 1187–1192 (1962).

[2]

M. Nikulin and F. Haghighi. On the power generalized Weibull family: model for cancer censored data. Metron – International Journal of Statistics 67, 75–86 (2009).