Reference

Index

SurvivalDistributions.ExpoDist

SurvivalDistributions.ExponentiatedWeibull

SurvivalDistributions.GeneralizedGamma

SurvivalDistributions.LogLogistic

SurvivalDistributions.PowerGeneralizedWeibull

SurvivalDistributions.censored_loglikelihood

SurvivalDistributions.censored_loglikelihood

SurvivalDistributions.cumhazard

SurvivalDistributions.cumhazard

SurvivalDistributions.hazard

SurvivalDistributions.hazard

SurvivalDistributions.loghazard

SurvivalDistributions.loghazard

SurvivalDistributions.ExpoDist — Type

ExpoDist(γ, X)

A power distribution with power γ and base distribution X<:ContinuousUnivariateDistribution is defined as the distribution that has cumulative distribution function $F^γ$ where F was the distribution function of X.

SurvivalDistributions.ExponentiatedWeibull — Type

ExponentiatedWeibull(α,θ,γ)

The Exponentiated Weibull distribution is obtain by exponentiating the cdf of the Weibull distribution. This simple transformation adds a second shape parameter that, interestingly, induces a lot of flexibility on the hazard function. The hazard function of the Exponentiated Weibull distribution can capture the basic shapes: constant, increasing, decreasing, bathtub, and unimodal, making it appealing for survival models.

A random variable X follows an ExponentiatedWeibull(α,θ,γ) distribution when it has cumulative distribution function $F_X = F_W^{γ}$ where $F_W$ is the cumulative distribution function of a Weibull(α,θ).

References:

Exponentiated Weibull distribution

Weibull distribution

SurvivalDistributions.GeneralizedGamma — Type

GeneralizedGamma(σ,nu,gamma)

The Generalised Gamma (GG) distribution is a three-parameter distribution with support on ${\mathbb R}_+$. The corresponding hazard function can accommodate bathtub, unimodal and monotone (increasing and decreasing) hazard shapes. The GG distribution has become popular in survival analysis due to its flexibility.

References:

Generalised Gamma

[1] Stacy, E.W. A generalization of the gamma distribution, The Annals of Mathematical Statistics, 1962

SurvivalDistributions.LogLogistic — Type

LogLogistic(μ,σ)

According to its wikipedia page, the the log-logistic distribution (known as the Fisk distribution in economics) is a continuous probability distribution for a non-negative random variable. It is used in survival analysis as a parametric model for events whose rate increases initially and decreases later, as, for example, mortality rate from cancer following diagnosis or treatment. It has also been used in hydrology to model stream flow and precipitation, in economics as a simple model of the distribution of wealth or income, and in networking to model the transmission times of data considering both the network and the software.

The log-logistic distribution is the probability distribution of a random variable whose logarithm has a logistic distribution. It is similar in shape to the log-normal distribution but has heavier tails. Unlike the log-normal, its cumulative distribution function can be written in closed form.

It is characterized by its density function as

\[f(x) = \frac{(\frac{β}{α})(\frac{x}{α})^{β-1} }{(1 + (\frac{x}{α})^{β})^2},\]

where α = e^μ and β = 1/σ.

SurvivalDistributions.PowerGeneralizedWeibull — Type

PowerGeneralizedWeibull(σ,ν,γ)

The Power Generalised Weibull (PGW) distribution is a three-parameter distribution with support on ${\mathbb R}_+$. The corresponding hazard function can accommodate bathtub, unimodal and monotone (increasing and decreasing) hazard shapes. The PGW distribution has become popular in survival analysis given the tractability of its hazard and survival functions.

The PowerGeneralizedWeibull(σ,ν,γ) distribution, with scale σ, shape ν (nu) and second shape γ has probability density function

\[f(t;σ,ν,γ) = \dfrac{ν}{γ σ^ν}t^{ν-1} \left[ 1 + \left(\dfrac{t}{σ}\right)^ν\right]^{\left(\frac{1}{γ}-1\right)} \exp\left\{ 1- \left[ 1 + \left(\dfrac{t}{σ}\right)^ν\right]^{\frac{1}{γ}} \right\}.\]

References:

[2] Nikulin, M. and Haghighi, F. On the power generalized Weibull family: model for cancer censored data. Metron – International Journal of Statistics, 2009

SurvivalDistributions.censored_loglikelihood — Method

censored_loglikelihood(X::UnivariateDistribution, t, δ)

Provide the censored logliklyhood of the distribution X at point t, with status indicatrix δ. Is if defined as

\[δ * loghazard(X,t) - cumhazard(X,t)\]

SurvivalDistributions.cumhazard — Method

loghazard(X::UnivariateDistribution, t)

Provide the cumulative hazard function of the random variable X. It is defined as $H(t) = - \ln S(t)$ where $S$ is the survival fucntion of the random variable X.

SurvivalDistributions.hazard — Method

hazard(X::UnivariateDistribution, t)

Provide the hazard fucntion of the random variable X (supposed to be a Distributions.ContinuousUnivariateDistributions) at point t. The default implementation is simply $vh(t) = \frac{f(t)}{S(t)}$ where $f$ and $S$ are the density and survival function of X.

SurvivalDistributions.loghazard — Method

loghazard(X, t)

Provide the log of the hazard fucntion of the random variable X. See hazard for the formal definition.

Bibliography

[1]

E. Stacy. A generalization of the gamma distribution. The Annals of Mathematical Statistics 33, 1187–1192 (1962).

[2]

M. Nikulin and F. Haghighi. On the power generalized Weibull family: model for cancer censored data. Metron – International Journal of Statistics 67, 75–86 (2009).